Reworked NodeTopologyEntity so that is used by enlinkd.
Removed enlinkd service Node
Need to change the TopologyCacheDao getAllTopologyEntity toi use OnmsCriteria

+++ tools/packages/sentinel/sentinel.spec
@@ -12,14 +12,14 @@
 %{!?sentinelrepoprefix:%define sentinelrepoprefix /opt/sentinel/repositories}
 
 # Description
-%{!?_name:%define _name "opennms"}
-%{!?_descr:%define _descr "OpenNMS"}
+%{!?_name:%define _name opennms}
+%{!?_descr:%define _descr OpenNMS}
 %{!?packagedir:%define packagedir %{_name}-%version-%{releasenumber}}
 
 %{!?_java:%define _java jre-1.8.0}
 
-%{!?extrainfo:%define extrainfo }
-%{!?extrainfo2:%define extrainfo2 }
+%{!?extrainfo:%define extrainfo %{nil}}
+%{!?extrainfo2:%define extrainfo2 %{nil}}
 %{!?skip_compile:%define skip_compile 0}
 %{!?enable_snapshots:%define enable_snapshots 1}
 
@@ -47,6 +47,9 @@ BuildArch:     noarch
 Source:        %{_name}-source-%{version}-%{releasenumber}.tar.gz
 URL:           http://www.opennms.org/wiki/Sentinel
 BuildRoot:     %{_tmppath}/%{name}-%{version}-root
+
+BuildRequires:	%{_java}
+BuildRequires:	libxslt
 
 #Requires:       jicmp >= 2.0.0
 #Requires(pre):  jicmp >= 2.0.0
@@ -94,6 +97,10 @@ rm -rf %{buildroot}
 export EXTRA_ARGS=""
 if [ "%{enable_snapshots}" = 1 ]; then
 	EXTRA_ARGS="-s"
+fi
+
+if [ "%{skip_compile}" = 1 ]; then
+	EXTRA_ARGS="$EXTRA_ARGS -c"
 fi
 
 tools/packages/sentinel/create-sentinel-assembly.sh $EXTRA_ARGS
@@ -157,7 +164,9 @@ ROOT_INST="${RPM_INSTALL_PREFIX0}"
 # Clean out the data directory
 if [ -d "${ROOT_INST}/data" ]; then
     find "$ROOT_INST/data/"* -maxdepth 0 -name tmp -prune -o -print0 | xargs -0 rm -rf
-    find "$ROOT_INST/data/tmp/"* -maxdepth 0 -name README -prune -o -print0 | xargs -0 rm -rf
+    if [ -d "${ROOT_INST}/data/tmp"  ]; then
+        find "$ROOT_INST/data/tmp/"* -maxdepth 0 -name README -prune -o -print0 | xargs -0 rm -rf
+    fi
 fi
 
 # Remove the directory used as the local Maven repo cache
